requirements.txt
langchain
langchain-google-genai
pandas
google-generativeai
